IOS-6059 Added SendAlertBuilder
